app/config.py
"""
Application configuration loaded from environment / .env file.
"""
import os
from dotenv import load_dotenv

load_dotenv()


class Config:
    OLLAMA_HOST: str = os.getenv("OLLAMA_HOST", "http://localhost:11434")
    DEEPSEEK_MODEL: str = os.getenv("DEEPSEEK_MODEL", "deepseek-r1:1.5b")
    CHATS_DIR: str = os.getenv("CHATS_DIR", "chats")
    FLASK_HOST: str = os.getenv("FLASK_HOST", "0.0.0.0")
    FLASK_PORT: int = int(os.getenv("FLASK_PORT", "5000"))
    FLASK_DEBUG: bool = os.getenv("FLASK_DEBUG", "false").lower() == "true"
