requirements.txt
flask==3.1.3
python-dotenv==1.2.2
ollama==0.6.1
pytest==9.0.3
